encode the modelfile
